clean up and consolidate caches
