+++ src/SemanticSimilarity/SemanticSimilarity.jl
@@ -2,14 +2,13 @@ module SemanticSimilarity
 
 using ..CrypticCrosswords: normalize
 
-# using SnowballStemmer: Stemmer, stem
 using WordNet
 using ProgressMeter: @showprogress
 using Statistics: mean
 
 export similarity, WuPalmer, SimilarityDepth
 
-include("caches.jl")
+include("cache.jl")
 include("metrics.jl")
 
 end # module
